LayoutTests/imported/w3c/web-platform-tests/domxpath/resolver-callback-interface-cross-realm-expected.txt

CONSOLE MESSAGE: TypeError: 'lookupNamespaceURI' property of callback interface should be callable
CONSOLE MESSAGE: TypeError: 'lookupNamespaceURI' property of callback interface should be callable
CONSOLE MESSAGE: TypeError: Proxy has already been revoked. No more operations are allowed to be performed on it
CONSOLE MESSAGE: TypeError: Proxy has already been revoked. No more operations are allowed to be performed on it
CONSOLE MESSAGE: TypeError: Proxy has already been revoked. No more operations are allowed to be performed on it


FAIL XPathNSResolver is cross-realm plain object without 'lookupNamespaceURI' property assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL XPathNSResolver is cross-realm plain object with non-callable 'lookupNamespaceURI' property assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL XPathNSResolver is cross-realm plain object with revoked Proxy as 'lookupNamespaceURI' property assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL XPathNSResolver is cross-realm non-callable revoked Proxy assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL XPathNSResolver is cross-realm callable revoked Proxy assert_unreached: Should have rejected: undefined Reached unreachable code
